DEV: Schema only dump for databases cache in `tests` workflow
The output from `pg_dumpall` is reused across tests workflow runs based
on the cache key of the `app-cache` actions cache. However, the dump
file contains database rows that have been inserted by our seed process.
The problem here is that our seed files may reference code in other
files but those files are not included as part of the cache key for
`app-cache`. It is also impractical for us to be able to identify all
possible files that alter the outcome of the seeding process.

As such, this commit updates the `tests` github actions workflow to only
dump the schema of the databases and will seed the databases from
scratch.