FIX: set category text color on field blur (#34289)
This change allows users to update category text color without
autocompleting the input value. Instead we are using `onBlur` to allow
the normalized hex value to be updated when the using submits or
tabs/clicks out of the field.

We are also restoring the normalized hex back to the correct value - for
example`f8f` becomes `ff88ff` if using only 3 digits. This was
mistakenly changed in #34218 to improve auto complete in the edit
category page but it affects the color pickers in other parts of the app
too.

---------

Co-authored-by: Jarek Radosz <jradosz@gmail.com>