UX: Move `selected` and `hover` colors up in the color palette UI (#34278)
The `selected` and `hover` colors are always positioned at the bottom of
the colors list in the admin color palette edit page. We think these
colors are more likely to be customized than many of the colors
currently shown above them, so this commit moves them a few places up so
they appear below the `header_primary` color.