FIX: Link to correct staff action logs for theme site setting
Fixes an issue where the history link on hover for a theme site setting
would point to the staff action logs for regular site settings, not  the
theme site setting ones which include the theme name in the subject
and also have a different `change_theme_site_setting` action.

c.f. https://meta.discourse.org/t/link-to-change-history-on-themeable-site-settings/377394