FIX: use group name in unreachable mention warning (#34174)
In cases where there are mentioned users that are unreachable (ie. they
cannot join the channel), then we normally notify the sender of those
users in the chat notice. However, with groups have a
`members_visible_level` setting which determines who can view members.
This change takes into account this setting when showing the chat notice
to the sender.

If there is a mentioned group that the sender does not have permission
to view members then we will show the group name instead of the users.