FEATURE: enable topic creation button for read only categories (#34211)
Most of this feature was already approved in #33495 but it was reverted
temporarily due to users accidentally posting in the wrong category (due
to the default composer category site setting).

This PR removes the default category from composer when the user doesn't
have permission to post in the category (preventing accidental posts in
the wrong category), and also inserts the read-only category into the
category chooser with a read only badge, this way providing some
information to the user when they find it in the dropdown.

The read only category in the drop down can't be selected via clicking
or keyboard selection.

<img width="1336" height="854" alt="Screenshot 2025-08-12 at 6 33 34 PM"
src="https://github.com/user-attachments/assets/a1640f5d-2c73-4c37-8f6a-e284982a8c12"
/>