DEV: Ensure that "confirm" user fields use boolean values (#34350)
This change fixes two issues with the "confirm" type user field.

These user fields are rendered as checkboxes, so the user field value
was being used directly as the `@checked` param for the input element.
Non-boolean values would cause unexpected checkbox behaviour (for
example, a value of `"false"` would cause the checkbox to be checked, or
a value of `null` would cause the form to be submitted with an empty
value, instead of `false`).

Additionally, since we know that these fields will only have boolean
values, `User#user_fields` now automatically casts the string `"true"`
and `"false"` values for confirm fields into their boolean equivalent.