DEV: simpler buildEngine used by prosemirror-editor (#34299)
This avoids messing with the `featuresOverride`, which we only did
before because of an attempt to filter the `onebox` feature out, which
may not be necessary.

The previous `buildEngine` call could strip information that's used by
customizations like `discourse-calendar`'s `[timezones]`, which relies
on the presence of a feature and hence is not being interpreted during
Markdown parsing as a token.