FIX: Refresh palettes list when navigating back to palettes list page (#34374)
Fixes
https://meta.discourse.org/t/checkbox-in-color-palettes-user-selectable-yes-no-doesnt-work/378545?u=osama

This commit moves the sorting logic to server-side to avoid caching issues
where the (sorted) palettes list, which was stored as a property on the
controller, would need to be invalidated each time the user navigated to
the palettes list page because changes to palettes could've been made
between the last and new visits to the page. This PR also eliminates an
AJAX request that the palettes list page makes to fetch the default
theme by including the default theme in the same response that includes
the color palettes list.