Remove duplicate entries from intersection additional tags (#34389)
Remove duplicate entries (or repeats of the main tag) from the
additional tags in the tags/intersection path.

This was causing unexpected behavior when the main tag was repeated
again in the additional tags (which isn't typically possible by the core
UI, but is possible in custom plugins or by directly editing the URL).

See https://meta.discourse.org/t/377740