FIX: Handle auxclick event on topic list (#34408)
This commit adds a new `auxclick` event handler to the topic list item,
which calls the same `click` handler.

The `auxclick` event is triggered by auxiliary mouse buttons, such as
the middle mouse button. This fixes an issue where using the middle
mouse button
on a topic card in Horizon anywhere but the topic title would not open
the topic
in a new tab.

Normal topic list click functionality remains unchanged.

c.f.
https://developer.mozilla.org/en-US/docs/Web/API/Element/auxclick_event