FIX: use correct rank value for abstention in voters query ordering
The query performed to fetch voters is ordered by digest, rank
and username, in that order. However, when ordering by rank,
the value 1 is used for abstention, but that is incorrect - it
should be 0 instead.

Because of the way the query is crafted, I believe it is impossible for this
bug to cause an issue, but I am unsure - I have not been able to
craft a test to reproduce a potential issue here.

c.f. https://meta.discourse.org/t/polls-voters-json-returning-duplicate-users-across-paged-requests/376636