UX: Do not alter post styling when displaying replies below
Simplify the class binding in the `post` component by always applying
`post__contents contents`.

This was a regression when modernizing the post-stream, although there is a condition to remove these classes in the widget code.
The state `repliesShown` was was never properly set in the old code.

This prevents styling issues when the embedded replies are displayed below the post