FIX: Required confirmations not showing up (stable) (#34508)
### What is the problem?

When adding a new field required "for all users", and the field is of "confirmation" type, it does not show up.

This is happening because we're checking for fields using `isEmpty`, assuming the value will be `null` before it is filled up, but for confirmation type fields it is in fact cast to `false`.

`isEmpty(value)` -> `!value`

This PR also fixes an issue where we'd "unlock" the user in the
front-end even if the call to update the required fields failed.