DEV: use welcome_banner.search placeholder in search banner (#34523)
We had a `welcome_banner.search` string, but it wasn't being used... so
it wasn't possible to replace the welcome banner placeholder
independently from other search inputs.

This adds the ability to pass in a custom placeholder for the
search-menu component so the text can be updated for individual cases as
needed.

Also added a test to ensure the translation is used.