DEV: Improve errors for failed impersonation (#34509)
I wanted to have a more specific error when there's an unexpected server error (5XX).

Turns out the existing error differentiation (between not_found and invalid) was not working, because the shape of e is not what the developer expected.

This PR fixes that and adds another message for unexpected errors.