FIX: removes support for include_expired param
This commit removes the `include_expired` param and its associated site setting `include_expired_events_on_calendar`.

Since recent changes this performance optimisation doesn’t make sense anymore as we only load the events needed for a specific before/after range.

The rules are the following:

Non recurring events:
- They have to have a start date between before/after range

Recurring events:
- they don't have event_dates, so we don't rely on it for them
- what's important for recurring_event, is the recurring_until property and the original starts at
- we want to return recurring events when their start date is is after the after param and the before param is before the recurring_until
- if recurring_until is NULL we just look at the start date

It's done this way as the serializer only sends the rrule to the front, and this is the frontend which will compute the recurring events, but if you want to have all the events of august, and your recurring event is starting in june, you need to also fetch this june event.

This commit ensures we can now deal properly with nil dates as it makes no sense to have dates for an expired recurring event.