FIX: in chat messages, filter uploads by UserUpload, not by Upload.user
When creating or updating a chat message with uploads,
use `UserUpload`s to filter uploads instead of only allowing
uploads by `Upload.user`. The latter is only the user that
originally created the upload, but any user with a valid `UserUpload`
should be able to reference an upload.

c.f. https://meta.discourse.org/t/chat-upload-bug/379253