DEV: Don't error out on invalid version strings to has_needed_version? (#34604)
If you pass an invalid version number to Discourse.has_needed_version? it raises an exception:

This is somewhat unexpected based on the method name, and interferes with conditionals.

I've searched around for usages, and all the call sites I can find do not handle this exception.

I also searched the logs across all our hosting, and there are no unhandled instances of this.

This PR proposes returning false on invalid string, and log an error.