DEV: Add capability to pass an icon setting into auth provider registration (#34584)
Add capability to pass an icon setting into auth provider registration to override the default icon with a site setting.

Matches the pattern we use for the auth provider title and pretty name. 

With this, the SAML plugin can be updated to have a site setting to override the default "user" icon.