UX: close autocomplete menu on space but allow for continued search of user full names (#34611)
The autocomplete menu currently takes too long to close after a
whitespace character is entered after an `@` mention word is typed in
the composer. This is due to the autocomplete waiting for a search to
complete as we allow for searching by full names for the mention
autocomplete feature in composers, and we want to keep the menu open
during the usual search cycle while typing the search term.

However, it is also reasonably common to add that space in order to
_stop_ autocompleting, so in order to handle this, we close the
autocomplete menu preemptively in such scenarios with mention
autocomplete and continue the search in the background. This is closer
to the older behaviour we had for closing the autocomplete menu upon
each subsequent search (which was removed because it looked like the
autocomplete menu was flickering).

Co-authored-by: Gary Pendergast <gary@pento.net>