FIX: New features git hash version tweaks (#34671)
There are two issues addressed by this PR:

We log an error to console when a git hash is passed to Discourse#has_needed_version?. This isn't correct. We should just return false.
We only check that the commit hash is in the repo, we don't check that it's actually an ancestor in the current tree. So you'd see the feature even if you're on a branch that doesn't have it.