FEATURE: Do not bump topic on post edit
We currently bump the entire topic if the last post
in the topic is edited.

This is quite noisy and distracting, particularly in cases where the
edit was to address a small typo rather than dramatically rework the
post.

So, this commit no longer bumps a topic on post edit for
any posts in the topic. Only adding non-staff tags to the
topic will bump it now.