FIX: Do not directly import `admin` module from main app (#34689)
This import will fail for non-admin users. Most of the time this isn't
causing a problem, since `edit-category-general` isn't imported by
non-admin logic. However, if a theme imports or calls `modifyClass` on
it, this import will be evaluated and fail.

This was affecting an initializer in the discourse-air theme for
non-admin users. However, the automatic error recovery for theme
initializers caught it, so the only impact was a console log.