FIX: Ensure ember version in cache key is coupled to compiler
During in-container updates, the old version of the application continues running while an update is applied. That means that it's possible for the `node_modules/ember-source` version to be different to the version currently loaded in the transpiler. That means it's theoretically possible for theme assets to be built with the old compiler, and then stored against the new version.

This commit removes that race condition by adding an `ember_version` method to the JS transpiler. This is guaranteed to give us an accurate version number for the template-compiler currently being used for themes.

This commit also bumps the BASE_COMPILER_VERSION to force a recompile on any sites affected by this race condition.