DEV: Fix a flaky signup system test
The test is flaky because the route which we are asserting for leads to
a redirect. If the assertion is called before the redirect, the test
passes. If the assertion is called after the redirect, the test fails.

Instead, we can just assert for the final state which we want to see
instead. The route itself is not important.