DEV: Fix a rare bug where topic scrolls to OP unexpectedly (#34713)
Repro steps:
- enable the assign plugin
- check the `unassigned on close` and `reassign on open` site settings
- assign a topic with more than 5 posts to a user
- view the topic and scroll to post 5 or more
- close the topic
- open the topic
- repeat a few times (behaviour can be inconsistent)

Expected: topic reloads without scrolling away from the current post.

Current behaviour: topic scrolls to OP. Notice in the video that it also
scrolls to OP for any other clients that are on the same topic
(left-side browser is of a random user).


https://github.com/user-attachments/assets/0cc79bf2-a2e9-4a1e-b3e3-5f2b19763d8c


Why? The assign plugin passes `reload_topic: true, refresh_stream: true`
via MessageBus. On the frontend code, when we reload the topic we did
not pass the `post_number` to the request, and therefore the current
position was lost. Passing the current post number to the topic reload
fixes the issue because then in the frontend the topic model maintains
the right position.

Another option is to not trigger `refresh_stream=true` here. It feels
unnecessary to do so for assignments that are opened/closed. That said,
other plugins also use this option and may have a valid reason to do so.