FEATURE: Hover menu for all chat channels in sidebar
This commit introduces a menu on an [...] button on hover
of DMs and public chat channels.

When you click on the menu you will see the following:

* Notification settings >
    * Send push notifications...
      * Never
      * Only for mentions
      * For all activity
    * Mute/unmute channel
* Leave channel
* Channel settings

Over time we may add more items to this menu, these are
just the most common actions for now.

This commit also moves the hovering functionality for sidebar
links to JS, in case we need to modify when the link should
stop hovering, e.g. if it's still inside the menu maybe we
don't want to.

Finally, I am introducing a `@suffixIcon` for DButton,
and a --text-only CSS class for generic headers for DropdownMenu
items