DEV: Remove unnecessary system test
This commit removes `spec/system/user_invites_spec.rb` added in
55bf0e21fb369a3f022c2c999337eeb10a870383. The system tests are flaky
because of the use of an arbitary `invite_redemption_delay` in
`User#invited_by`. In system test, the user is always going to be
created before the delay.

Also I don't think a system test is required here since we have unit
tests for the `User#invited_by` method. Hence, I'm removing the system
test instead of just fixing it.