FIX: Define the typical AFL Grand Final date. (#34768)
The AFL Grand Final is typically help on the last Saturday of September,
though there are occasional exceptions. Since 2015, the Friday before
the Grand Final is a public holiday in Victoria, Australia.

This change defines the typical date for the public holiday, but allows
exceptions to be defined where necessary.