PERF: do not fetch thread data when we have it (#34799)
Prior to this fix we would fetch thread data even if it was present in
the bus data, that's already unnecessary work in a normal situation but
that's even worse when a user is coming back to their computer as we
will have to catchup multiple new threads at once.

Also ensures we correctly await fetching data from server when
necessary.

No test as there are not behavior change.