FIX: ensures event serializer has tags info (#34801)
This was preventing the `map_events_to_color` site setting to work
correctly given we couldn't match on tag slug as it was not present.

This commit adds system specs and removes the js test as it was
misleading and incorrectly passing due to the fixture which couldn't
catch the issue in the serializer.