FIX: reset hashtags when pressing return in composer (#34579)
Fixes an issue where a hashtag is decorated in composer and we try to
either:

- press return in middle of a paragraph to move the hashtag onto the
next line
- press backspace at beginning of a paragraph containing emojis

In both scenarios the hashtag decoration was lost but the data
attributes were retained (ie. `processed` was set to `true`). By setting
`processed` to `false` we can force hashtags to be processed again after
breaking into a new paragraph or when pressing backspace at the start of
the line to combine two paragraphs.