FEATURE: new topic button always enabled (#34603)
### What is this change?

When visiting a restricted category or tag where the current user does
not have permission to create a topic, we should still allow the user to
write a new topic under a different category or tag. This is especially
important when using the new topic button from the sidebar.

#### When the user does not have permission, the composer category
defaults to:

1) Subcategory (if SiteSetting.default_subcategory_on_read_only_category
is true and subcategory exists)
2) Default category chooser value of Category... prompting the user to
select a category from the dropdown

This PR is a follow up to #33495 which was reverted previously due to a
couple of issues which are now corrected:

- we no longer set a default category based on
`default_composer_category` (previously was step 2 above)
- category banner text when a user cannot create a topic is now showing
up correctly
- added a title attribute with more info for the disabled category in
the category dropdown in composer
- for staff only tags, when creating a topic from the tag page as a non
staff member we now automatically remove the restricted tag from the
tags field in composer (as the user cannot use this tag)