FIX: `applySurround` should detect both sides when deleting surrounde… (#34243)
…d text

In the current Textarea Editor `applySurround`, only the right side is
checked. That means if you write `something**`, pressing `Ctrl + B` will
unexpectedly turn it into `mething`, since the
`TextareaTextManipulation` detected the double asterisk `**` on the
right and thought `something**` was the bold syntax, so it removed it.

This commit fixes the bug here.

Related meta topic: https://meta.discourse.org/t/-/358800