DEV: suppress noisy warning from system test runs (#34855)
The halfvec type is not recognized by Rails, this is not normally an
issue
cause we do not use Active Record to query this anyway.

However a workaround we have in place adds all data_sources to
schema_cache
forcing the PG gem to try to resolve the type.

This workaround suppresses this specific warning on STDERR so we don't
infect
our test results.

This is particularly confusing to AI agents that see the text:

"failed to recognize type of ..." which leads them to think a system
spec run
failed.