FIX: DiscourseConnect provider redirect to forum instead of the sso_redirect_url in the payload (#34580)
**Description**

Currently, when using Discourse as an SSO provider through the
DiscourseConnect provider, if a user logs in using the login form, the
user is not redirected to the expected `return_sso_url`. Instead, it is
redirected to the forum's home page. This fixes this behaviour and
allows the user to return to the expected URL.