DEV: rich editor - don't parse html images with `data:` content (#34863)
Avoid parsing images with `data:` content (usually with long base64) on
image node's `parseDOM` (eg. when pasting).

Base64 images are unsupported when pasting on markdown mode and on other
"rich" authoring tools.

A future improvement will be triggering an upload for each pasted base64
image and then replacing it, but this stops the bleeding of generating
an awfully long Markdown output.