FIX: Horizon topic activity username sometimes blank (#34679)
For these steps:

* In Horizon, create a topic.
* You’ll notice that the topic card has the OP name.
* Make an edit to the OP after the edit grace period.

You would see no username under the topic title. This is because
we were saying that if the topic bump date was after the last post
date, then we should never show any edit details.

However this didn't handle the common case of posting then editing
soon after.

Now in this commit, if the difference is < 1 day, we still show
the same username.

A proper fix would be to know the bumped_at_user then we can tell
if the user is the same as the last poster, but that is more complex
to do.