DEV: Allow `ServerSession` to store arbitrary data
Currently, the server session can only store strings. As we want to move
more things into it (like with a proper session object), we need to be
able to store arbitrary data.

This patch serializes data using Message Pack, as it’s almost as
flexible as Marshal but without the potential security issues.