FIX: Do not bump topics on title, category, or tag edit
Followup 2a65bf452224ece92981a9a26579ece67f10855b

With the above change, we inadvertantly made it so that
when you edited the topic title or category, we bumped
the topic.

This fix makes it so we do not bump topics when only title, category,
or tags are changed, to restore the previous behaviour.