DEV: Improve system theme loading
Use `set_field` and `theme.save!` when updating the theme. This ensures all the correct caches are rebuilt, just like when you change a file via the admin UI. It also sends more precise refresh commands to the browser so that CSS-only changes apply live, instead of requiring a full page refresh.