DEV: Improve rspec boot time & `--bisect` functionality
By default, RSpec's bisect uses the `:fork` strategy. This is not safe to use in Discourse, because miniracer is not fork safe. So if any contexts have been booted in the parent process, they'll be broken in the forked process. Using `:shell` solves this.

Also, since we introduced core themes, the database 'seed' step became much more expensive because it re-installs and compiles Horizon/Foundation every time. I don't think there's any need to do this every time we boot a spec. Instead, we can trust that it ran as part of `bin/rake db:migrate`, just like it does in production.

On my machine, this cuts the runtime of a no-op `bin/rspec` from 3.5s to 1.1s 🚀