DEV: Move convert video job to upload after create
This moves the logic for creating the convert_video job to the upload
after_create hook so that we ensure there is an upload_id. When this logic
was in the s3 store and direct to s3 uploads was enabled the convert video job
would never fire because we didn't have an upload_id.