DEV: Properly clean up global setting overrides in specs
In the `SiteSettingExtension`, when calling `SiteSetting.send(:setting_name, value)`,
if the setting counts as a global one, we add it to the list of hidden
and shadowed site settings in memory. However, when using the
`global_setting` helper in specs, we were not cleaning up these
overrides after the spec finished, leading to potential flakiness
in tests that rely on the visibility of site settings.

The fix is to just remove the setting from the hidden and shadowed
list after the spec is done.